Creating components PaymentComponent and Products
