--- src/components/TabsComponent/index.css
@@ -1,3 +1,0 @@
-.MuiTab-textColorPrimary-142.MuiTab-selected-144 {
-  color: #FF7800 !important
-}
