--- src/screens/home/index.css
@@ -1,3 +1,0 @@
-/* .MuiTab-textColorPrimary-139.MuiTab-selected-141 {
-  color: #FF7800 !important
-} */
